Move updatable hover interfaces to hover.ts
